Switch from bpf_redirect_lookup to bpf_map_lookup + bpf_sk_assign
After we have enabled socket lookup in sockmap, the approach for
selecting a socket has changed. Instead of calling a helper that
takes a SOCKMAP and an index into it as argument, lookup up a socket
in SOCKMAP, and select it with bpf_sk_assign helper.

Adapt our connection dispatching program to use the new approach.